Refactors register page - moves user and pass into Account
